python/examples/crewai/requirements.txt
lightspark
crewai
python-dotenv
